Working page load actions computation. TODO : Update the tests for correctness.
